pyproject.toml
[tool.poetry]
name = "videocolorsbarcode"
version = "0.1.0"
description = "Generates a colors barcode by plotting the average color of each frame of a video."
authors = ["Federico Lanzani <federico@federicolanzani.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.21.0"
opencv-python-headless = "^4.7.0.72"
colorthief = "^0.2.1"
tqdm = "^4.65.0"
pillow = "^9.5.0"
pandas = "^1.5.2"
numpy = "^1.24.2"




[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
